TR-1: Changed architecture and moved steps
