--- TestRailTestingJS/src/test/logout.js
@@ -1,16 +1,0 @@
-const { config } = require("../../wdio.conf")
-const LoginPage = require('../pageobjects/login.page');
-const HomePage = require('../pageobjects/home.page');
-
-describe('My Logout application', () => {
-    it('should be successful', async () => {
-        await LoginPage.open();
-        await LoginPage.login(config.userData.email, config.userData.password);
-
-        await HomePage.logout();
-
-        await expect(browser).toHaveUrl(`${config.baseUrl}index.php?/auth/login`);
-        await expect(browser).toHaveTitle('Login - TestRail');
-        await expect(LoginPage.inputUsername).toBePresent;
-    });
-});
