+++ package.json
@@ -14,11 +14,13 @@
   },
   "devDependencies": {
     "@wdio/allure-reporter": "^7.16.14",
-    "@wdio/cli": "^7.16.14",
-    "@wdio/local-runner": "^7.16.14",
-    "@wdio/mocha-framework": "^7.16.14",
+    "@wdio/cli": "^7.16.15",
+    "@wdio/local-runner": "^7.16.15",
+    "@wdio/mocha-framework": "^7.16.15",
     "@wdio/spec-reporter": "^7.16.14",
     "chromedriver": "^97.0.4",
+    "ts-node": "^10.5.0",
+    "typescript": "^4.5.5",
     "wdio-chromedriver-service": "^7.2.6"
   }
 }
