+++ src/pages/home.page.js
@@ -1,18 +1,10 @@
-const Page = require('./page');
-
-class HomePage extends Page {
-
+class HomePage {
     get spanText() {
         return $('.navigation-username');
     }
 
     get hrefLogOut() {
         return $('#navigation-user-logout');
-    }
-
-    async logout() {
-        await this.spanText.click();
-        await this.hrefLogOut.click();
     }
 }
