+++ src/pages/login.page.js
@@ -1,0 +1,15 @@
+class LoginPage {
+    get inputUsername() {
+        return $('#name');
+    }
+
+    get inputPassword() {
+        return $('#password');
+    }
+
+    get btnSubmit() {
+        return $('#button_primary');
+    }
+}
+
+module.exports = new LoginPage();
