+++ src/steps/base.step.js
@@ -1,6 +1,6 @@
 const { config } = require("../../wdio.conf")
 
-module.exports = class Page {
+module.exports = class BaseStep {
     open() {
         return browser.url(config.baseUrl);
     }
