+++ src/steps/logout.step.js
@@ -1,0 +1,11 @@
+const BaseStep = require('./base.step');
+const HomePage = require('../pages/home.page');
+
+class LogoutStep extends BaseStep {
+    async logout() {
+        await HomePage.spanText.click();
+        await HomePage.hrefLogOut.click();
+    }
+}
+
+module.exports = new LogoutStep();
