+++ src/tests/login.js
@@ -1,12 +1,12 @@
 const { config } = require("../../wdio.conf")
-const LoginPage = require('../pageobjects/login.page');
-const HomePage = require('../pageobjects/home.page');
+const LoginStep = require('../steps/login.step');
+const HomePage = require('../pages/home.page');
 
 describe('My Login application', () => {
     it('should login with valid credentials', async () => {
-        await LoginPage.open();
+        await LoginStep.open();
 
-        await LoginPage.login(config.userData.email, config.userData.password);
+        await LoginStep.login(config.userData.email, config.userData.password);
 
         await expect(HomePage.spanText).toHaveTextContaining(config.userData.fullName);
         await expect(browser).toHaveUrlContaining('dashboard');
